refactor: Rename file `variables.py` to `variable_loader` for clarity
