+++ aws_annoying/variable_loader.py
